Issue #151 Use shorts for indexing arrays in all benchmarks. (Note: FFT already used shorts, so no change there)

+++ src/lib/bm_binsrchX/java/javax/rtcbench/RTCBenchmark.java
@@ -32,15 +32,15 @@ public class RTCBenchmark {
         short toFind = (short)(numbers[0] - 1);
 
         for (short i=0; i<1000; i++) {
-            int low = 0;
-            int high = numbers.length - 1;
-            int mid;
+            short low = 0;
+            short high = (short)(numbers.length - 1);
+            short mid;
             while (low <= high) {
-                mid = (low + high) / 2;
+                mid = (short)((low + high) / 2);
                 if (numbers[mid] < toFind) {
-                    low = mid + 1;
+                    low = (short)(mid + 1);
                 } else if (numbers[mid] > toFind) {
-                    high = mid - 1;
+                    high = (short)(mid - 1);
                 } else {
                     break; // Found. Would return from here in a normal search, but for this benchmark we just want to try many numbers.
                 }
