+++ src/lib/bm_sorto/java/javax/rtcbench/RTCBenchmark.java
@@ -46,10 +46,10 @@ public class RTCBenchmark {
     public static void rtcbenchmark_measure_java_performance(short[] numbers) {
         Stopwatch.resetAndStart();
 
-        for (int i=0; i<NUMNUMBERS; i++) {
-            int x=(NUMNUMBERS-i-1); // This doesn't get optimised the way I expected it would. Without this extra variable, it will calculate NUMNUMBERS-i-1 on each interation of the inner loop! (speedup 14.7M -> 14.2M cycles)
-            int j_plus_one = 1; // Same goes for "j+1"
-            for (int j=0; j<x; j++) {
+        for (short i=0; i<NUMNUMBERS; i++) {
+            short x=(short)(NUMNUMBERS-i-1); // This doesn't get optimised the way I expected it would. Without this extra variable, it will calculate NUMNUMBERS-i-1 on each interation of the inner loop! (speedup 14.7M -> 14.2M cycles)
+            short j_plus_one = 1; // Same goes for "j+1"
+            for (short j=0; j<x; j++) {
                 short val_at_j = numbers[j];
                 short val_at_j_plus_one = numbers[j_plus_one];
                 if (val_at_j>val_at_j_plus_one) {
